+++ demo/bridge/cachecache/index.php
@@ -1,13 +1,6 @@
 <?php
 
-include __DIR__ . '/vendor/autoload.php';
-include __DIR__ . '/../../bootstrap.php';
-
-$debugbarRenderer->setBaseUrl('../../../src/DebugBar/Resources');
-
-$cache = new CacheCache\Cache(new CacheCache\Backends\Memory());
-
-$debugbar->addCollector(new DebugBar\Bridge\CacheCacheCollector($cache));
+include __DIR__ . '/bootstrap.php';
 
 $cache->set('foo', 'bar');
 $cache->get('foo');